+++ mayan/apps/user_management/apps.py
@@ -10,7 +10,7 @@ from mayan.apps.acls.links import link_acl_list
 from mayan.apps.acls.permissions import permission_acl_edit, permission_acl_view
 from mayan.apps.common.apps import MayanAppConfig
 from mayan.apps.common.menus import (
-    menu_multi_item, menu_object, menu_secondary, menu_setup
+    menu_multi_item, menu_object, menu_secondary, menu_setup, menu_user
 )
 from mayan.apps.common.widgets import TwoStateWidget
 from mayan.apps.metadata import MetadataLookup
@@ -19,11 +19,13 @@ from mayan.apps.rest_api.fields import DynamicSerializerField
 
 from .handlers import handler_initialize_new_user_options
 from .links import (
-    link_group_create, link_group_delete, link_group_edit, link_group_list,
-    link_group_members, link_group_setup, link_user_create, link_user_delete,
-    link_user_edit, link_user_groups, link_user_list,
-    link_user_multiple_delete, link_user_multiple_set_password,
-    link_user_set_options, link_user_set_password, link_user_setup
+    link_current_user_details, link_current_user_edit, link_group_create,
+    link_group_delete, link_group_edit, link_group_list, link_group_members,
+    link_group_setup, link_user_create, link_user_delete, link_user_edit,
+    link_user_groups, link_user_list, link_user_multiple_delete,
+    link_user_multiple_set_password, link_user_set_options,
+    link_user_set_password, link_user_setup, separator_user_label,
+    text_user_label
 )
 from .permissions import (
     permission_group_delete, permission_group_edit,
@@ -145,6 +147,12 @@ class UserManagementApp(MayanAppConfig):
             )
         )
         menu_setup.bind_links(links=(link_user_setup, link_group_setup))
+        menu_user.bind_links(
+            links=(
+                text_user_label, separator_user_label,
+                link_current_user_details, link_current_user_edit,
+            ), position=0
+        )
 
         post_save.connect(
             dispatch_uid='user_management_handler_initialize_new_user_options',