+++ mayan/apps/user_management/forms.py
@@ -5,6 +5,9 @@ from django.contrib.auth import get_user_model
 
 
 class UserForm(forms.ModelForm):
+    """
+    Form used to edit an user's mininal fields by the user himself
+    """
     class Meta:
         fields = (
             'username', 'first_name', 'last_name', 'email', 'is_active',