+++ mayan/apps/user_management/icons.py
@@ -2,6 +2,8 @@ from __future__ import absolute_import, unicode_literals
 
 from mayan.apps.appearance.classes import Icon
 
+icon_current_user_details = Icon(driver_name='fontawesome', symbol='user')
+icon_current_user_edit = Icon(driver_name='fontawesome', symbol='user')
 icon_group_create = Icon(driver_name='fontawesome', symbol='plus')
 icon_group_setup = Icon(driver_name='fontawesome', symbol='users')
 icon_user_create = Icon(driver_name='fontawesome', symbol='plus')